Move SinglePhotosView out into it's own swift file
